Store currency as int instead of float

+++ backend/src/database/user.py
@@ -10,7 +10,7 @@ class User(Base, SerializerMixin):
     card_uid = Column(String, nullable=False)
     creation_date = Column(DateTime, default=datetime.now)
     update_date = Column(DateTime, default=None)
-    currency_amount = Column(Float, default=0)
+    currency_amount = Column(Integer, default=0)
     name = Column(String, default=None)
     first_name = Column(String, default=None)
     email = Column(String, default=None)
